+++ NEWS.md
@@ -1,3 +1,4 @@
+# Changes in v1.0.2
 
 * Add the `auto_weight` argument to `textmodel_lss()` and `as.textmodel_lss()` to improve the accuracy of scaling.
 * Remove the `group` argument from `textplot_simil()` to simplify the object.